# Conflicts:
#	src/main/java/com/jwt/controllers/KafkaTest.java

+++ src/main/java/com/jwt/security/server/PortServiceIpml.java
@@ -2,6 +2,7 @@ package com.jwt.security.server;
 
 import com.jwt.models.CustomerPort;
 import com.jwt.repository.PortsRepository;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.apache.catalina.LifecycleException;
 import org.springframework.http.ResponseEntity;
@@ -18,12 +19,18 @@ public class PortServiceIpml implements PortService {
 
   private final PortServiceListener portServiceListener;
 
+  private final HttpServletRequest request;
+
   @Override
   public ResponseEntity<?> portRegister(@RequestBody CustomerPort port) throws LifecycleException {
     Optional<CustomerPort> ports = portsRepository.findByPort(port.getPort());
+    int adminPort = request.getLocalPort();
     if (ports.isPresent()) {
       System.out.println("Port is existed !");
       return ResponseEntity.badRequest().body("Port is existed !");
+    } else if (adminPort != 8082) {
+      System.out.println("You do not have permission to do this !");
+      return ResponseEntity.badRequest().body("You do not have permission to do this !");
     } else {
       portsRepository.save(port);
       portServiceListener.addPort(port.getPort());
@@ -34,10 +41,15 @@ public class PortServiceIpml implements PortService {
 
   @Override
   public ResponseEntity<?> deletePort(@RequestBody CustomerPort port) throws LifecycleException {
-    portsRepository.deletePort(port.getPort());
-    portServiceListener.removePort(port.getPort());
-    System.out.println("Successfully destroy port: " + port.getPort());
-    return ResponseEntity.ok("Successfully destroy port: " + port.getPort());
+    int adminPort = request.getLocalPort();
+    if(adminPort == 8082){
+      portsRepository.deletePort(port.getPort());
+      portServiceListener.removePort(port.getPort());
+      System.out.println("Successfully destroy port: " + port.getPort());
+      return ResponseEntity.ok("Successfully destroy port: " + port.getPort());
+    } else {
+      System.out.println("You do not have permission to do this !");
+      return ResponseEntity.badRequest().body("You do not have permission to do this !");
+    }
   }
 }
-
